+++ protected/humhub/views/layouts/main.php
@@ -25,58 +25,20 @@ AppAsset::register($this);
         <?php $this->head() ?>
 
         <!-- The HTML5 shim, for IE6-8 support of HTML5 elements -->
-        <!--[if lt IE 9]>
-        <script src="<?php echo Yii::getAlias(" @web"); ?>/js/html5shiv.js"></script>
-        <
-        link
-        id = "ie-style"
-        href = "<?php echo Yii::getAlias("
-        @
-        web
-        "); ?>/css/ie.css"
-        rel = "stylesheet" >
-        <![endif]-->
 
+        <!--[if lt IE 9]>
+        <script src="<?php echo Yii::getAlias("@web"); ?>/js/html5shiv.js"></script>
+        <link id="ie-style" href="<?php echo Yii::getAlias("@web"); ?>/css/ie.css" rel="stylesheet">
+        <![endif]-->
         <!--[if IE 9]>
-        <link id="ie9style" href="<?php echo Yii::getAlias(" @web"); ?>/css/ie9.css" rel="stylesheet">
+        <link id="ie9style" href="<?php echo Yii::getAlias("@web"); ?>/css/ie9.css" rel="stylesheet">
         <![endif]-->
 
         <!-- start: render additional head (css and js files) -->
         <?php echo $this->render('head'); ?>
         <!-- end: render additional head -->
-
-
-        <!-- start: Favicon and Touch Icons -->
-        <link rel="apple-touch-icon" sizes="57x57" href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-57x57.png">
-        <link rel="apple-touch-icon" sizes="60x60" href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-60x60.png">
-        <link rel="apple-touch-icon" sizes="72x72"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-72x72.png">
-        <link rel="apple-touch-icon" sizes="76x76" href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-76x76.png">
-        <link rel="apple-touch-icon" sizes="114x114"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-114x114.png">
-        <link rel="apple-touch-icon" sizes="120x120"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-120x120.png">
-        <link rel="apple-touch-icon" sizes="144x144"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-144x144.png">
-        <link rel="apple-touch-icon" sizes="152x152"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-152x152.png">
-        <link rel="apple-touch-icon" sizes="180x180"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/apple-icon-180x180.png">
-        <link rel="icon" type="image/png" sizes="192x192"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/android-icon-192x192.png">
-        <link rel="icon" type="image/png" sizes="32x32"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/favicon-32x32.png">
-        <link rel="icon" type="image/png" sizes="96x96"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/favicon-96x96.png">
-        <link rel="icon" type="image/png" sizes="16x16"
-              href="<?php echo Yii::getAlias("@web"); ?>/ico/favicon-16x16.png">
-        <link rel="manifest" href="<?php echo Yii::getAlias("@web"); ?>/ico/manifest.json">
-        <meta name="msapplication-TileColor" content="#ffffff">
-        <meta name="msapplication-TileImage" content="<?php echo Yii::getAlias("@web"); ?>/ico/ms-icon-144x144.png">
-        <meta name="theme-color" content="#ffffff">
+        
         <meta charset="<?= Yii::$app->charset ?>">
-        <!-- end: Favicon and Touch Icons -->
-
     </head>
 
     <body>
@@ -95,13 +57,11 @@ AppAsset::register($this);
                 </div>
 
                 <div class="notifications pull-right">
-
                     <?php
                     echo \humhub\widgets\NotificationArea::widget(['widgets' => [
                             [\humhub\modules\notification\widgets\Overview::className(), [], ['sortOrder' => 10]],
                     ]]);
                     ?>
-
                 </div>
 
             </div>