+++ protected/humhub/docs/CHANGELOG.md
@@ -3,6 +3,6 @@ HumHub Change Log
 
 1.2.0 under development
 -----------------------
-
+- Enh: Moved HumHub browser icons to HumHub theme (luke-)